backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
django-channels==4.0.0
channels-redis==4.1.0